Invoice send receive

+++ app/backend/src/main/java/org/letspeppol/app/dto/MaxProcessedDto.java
@@ -1,5 +1,5 @@
 package org.letspeppol.app.dto;
 
 public record MaxProcessedDto(
-    Integer maxDailyTotal
+    Long maxDailyTotal
 ) {}
